Change header link gap size
Increased the gap size between the header links in order
to improve the visual distinction between links. This was
necessary because one of the links contains two words with
a space in between them, which can cause the two words to
be interpreted as two separate links due to the space and gap
being a similar size.